Local Async Storage of Classes and Grades
Unless the user swipes down in the gradebook screen to manually reload, the classes and grades will only be fetched once when loading the app and will persist beyond the focus on the gradebook screen. This is done to decrease server fetch dependencies (and thus loading times) as well as to allow other screens to more efficiently be able pull classes and grades data locally.