Add: SwipeableItem, Fix: delete event bug
SwipeableItem: swipe left from the right of each item/event in whatever section to see the edit icon, press to open edit modal.

Fixed a bug where after you delete the last item of an open section, the section does not close, causing it to not open when a new item is added in that same section.