Add: selectable calendar dates, Fix: formatting bugs, code restructuring
The 'day' property of selectedDate and displayDate objects is the day of the month. Not to be confused with Date.prototype.getDay() which gets the day of the week.

Also, all dates and related variables stored that are not of the Javascript Date class have 1-indexed months, i.e. months in the range 1 to 12, inclusive.

Note the distinction between displayDate and selectedDate as commented in the code.