Add: swipe right on event to set its priority, Fix: minor bugs
I could not figure out how to get the value of the parameter percentOpen (of type  Animated.Node<number>) in function renderUnderlayRight. In turn, I was not able to check that value to get it to automatically set the priority as soon as a certain percent is reached, so you have to swipe then tap to set its priority. This is probably because I cannot control the code that is run during Animated.View re-renders.