Add: planner event due time + upgraded React Native version, Fix: iOS building
Upgraded React Native version to 0.65.1 to allow for successful iOS building. This may cause issues with Reanimated 2 (as it is not compatible with the upgraded React version), and thus Android building may not work. TODO: upgrade Reanimated 2 (and React Native Gesture Handler?).

Bugs for iOS include z-index not working (for planner dropdown menus), intral-server not connecting, textAlignVertical not working, planner event section nested scrolling, and stack screen switching to profile page when clicking anywhere but a button in the planner page.